+++ src/constants.h
@@ -1,0 +1,13 @@
+#ifndef CONSTANTS_H
+#define CONSTANTS_H
+
+#include <curses.h>
+
+// Server parameters
+int map_height = 80;
+int map_width = 160;
+int map_panel_height = 20;
+int map_panel_width = 40;
+
+
+#endif /* CONSTANTS_H */