+++ docs/HISTORY.md
@@ -5,5 +5,5 @@ Changelog
 Alpha (unreleased)
 ------------------
 
-- Constructed curses pad view to represent map panel
+- Constructed curses pad view to represent map panel  
   [Shaun Marshall]